Work on initialising simulations -- setting initial conditions etc

+++ src/ibis/src/commands/clean.py
@@ -1,11 +1,13 @@
 import shutil
+import os
 
 from ibis_py_utils import read_defaults
 
 def main():
     directories = read_defaults("directories.json") 
-    for dir in ("config_dir", "grid_dir"):
-        try:
+    for dir in ("config_dir", "grid_dir", "flow_dir"):
+        if os.path.exists(dir):
             shutil.rmtree(directories[dir])
-        except:
-            pass
+
+    if os.path.exists("__pycache__"):
+        shutil.rmtree("__pycache__")
